Add files via upload
I have created an IPL Cricket Analysis Report using Power BI. Check it out! The report provides a comprehensive overview of the Indian Premier League, its performance analysis over the years, and its team performance. It also compares the performance of different teams and players in the IPL. The report makes use of various interactive visuals to make the data easier to understand and interpret.